# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "Accepted": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/JobOut"
          }
        },
        "required": [
          "job"
        ],
        "title": "Accepted",
        "type": "object"
      },
...
      "DashboardSummary": {
        "properties": {
          "active_jobs": {
            "title": "Active Jobs",
            "type": "integer"
          },
          "instances": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "integer"
              },
              "type": "object"
            },
            "title": "Instances",
            "type": "object"
          },
          "projects": {
            "title": "Projects",
            "type": "integer"
          },
          "recent_jobs": {
            "items": {
              "$ref": "#/components/schemas/JobOut"
            },
            "title": "Recent Jobs",
            "type": "array"
          }
        },
        "required": [
          "projects",
          "instances",
          "active_jobs",
          "recent_jobs"
        ],
        "title": "DashboardSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IPv4Config": {
        "additionalProperties": false,
        "description": "Static IPv4 typed by the user (no IPAM in the MVP).",
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          },
          "dns": {
            "default": [],
            "items": {
              "type": "string"
            },
            "maxItems": 3,
            "title": "Dns",
            "type": "array"
          },
          "gateway": {
            "title": "Gateway",
            "type": "string"
          }
        },
        "required": [
          "address",
          "gateway"
        ],
        "title": "IPv4Config",
        "type": "object"
      },
...
      "InstanceAccepted": {
        "properties": {
          "instance": {
            "$ref": "#/components/schemas/InstanceOut"
          },
          "job": {
            "$ref": "#/components/schemas/JobOut"
          }
        },
        "required": [
          "job",
          "instance"
        ],
        "title": "InstanceAccepted",
        "type": "object"
      },
      "InstanceCreate": {
        "additionalProperties": false,
        "properties": {
          "image_id": {
            "format": "uuid",
            "title": "Image Id",
            "type": "string"
          },
          "ipv4": {
            "$ref": "#/components/schemas/IPv4Config"
          },
          "memory_mb": {
            "maximum": 262144.0,
            "minimum": 512.0,
            "multipleOf": 256.0,
            "title": "Memory Mb",
            "type": "integer"
          },
          "name": {
            "pattern": "^[a-z]([a-z0-9-]{0,61}[a-z0-9])?$",
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "root_disk_gb": {
            "maximum": 2048.0,
            "minimum": 1.0,
            "title": "Root Disk Gb",
            "type": "integer"
          },
          "ssh_key_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 10,
            "minItems": 1,
            "title": "Ssh Key Ids",
            "type": "array"
          },
          "tags": {
            "default": [],
            "items": {
              "pattern": "^[a-z0-9][a-z0-9_.-]{0,31}$",
              "type": "string"
            },
            "maxItems": 10,
            "title": "Tags",
            "type": "array"
          },
          "vcpus": {
            "maximum": 32.0,
            "minimum": 1.0,
            "title": "Vcpus",
            "type": "integer"
          },
          "zone_id": {
            "format": "uuid",
            "title": "Zone Id",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "name",
          "image_id",
          "zone_id",
          "vcpus",
          "memory_mb",
          "root_disk_gb",
          "ssh_key_ids",
          "ipv4"
        ],
        "title": "InstanceCreate",
        "type": "object"
      },
      "InstanceDelete": {
        "additionalProperties": false,
        "properties": {
          "confirm": {
            "title": "Confirm",
            "type": "string"
          }
        },
        "required": [
          "confirm"
        ],
        "title": "InstanceDelete",
        "type": "object"
      },
      "InstanceOut": {
        "description": "Tenant view. No provider identifiers (vmid/node/cluster) — ADR-0010.",
        "properties": {
          "cpu_usage": {
            "title": "Cpu Usage",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "gateway": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gateway"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Id"
          },
          "ipv4": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ipv4"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "memory_mb": {
            "title": "Memory Mb",
            "type": "integer"
          },
          "memory_used_mb": {
            "title": "Memory Used Mb",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "power_state": {
            "title": "Power State",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "region_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Name"
          },
          "root_disk_gb": {
            "title": "Root Disk Gb",
            "type": "integer"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "uptime_seconds": {
            "title": "Uptime Seconds",
            "type": "integer"
          },
          "vcpus": {
            "title": "Vcpus",
            "type": "integer"
          },
          "zone_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone Id"
          },
          "zone_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone Name"
          }
        },
        "required": [
          "id",
          "project_id",
          "kind",
          "name",
          "state",
          "power_state",
          "vcpus",
          "memory_mb",
          "root_disk_gb",
          "tags",
          "image_id",
          "zone_id",
          "zone_name",
          "region_name",
          "cpu_usage",
          "memory_used_mb",
          "uptime_seconds",
          "ipv4",
          "gateway",
          "created_at",
          "last_seen_at"
        ],
        "title": "InstanceOut",
        "type": "object"
      },
...
      "Page_InstanceOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InstanceOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[InstanceOut]",
        "type": "object"
      },
      "Page_JobOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[JobOut]",
        "type": "object"
      },
...
      "QuotaLineOut": {
        "properties": {
          "available": {
            "title": "Available",
            "type": "integer"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "resource": {
            "title": "Resource",
            "type": "string"
          },
          "used": {
            "title": "Used",
            "type": "integer"
          }
        },
        "required": [
          "resource",
          "limit",
          "used",
          "available"
        ],
        "title": "QuotaLineOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Tenant-scoped routes (projects, instances, jobs, role bindings, dashboard) require the header `X-Tenant-Id: <tenant uuid>`; it is validated against the caller's memberships. It is omitted from each operation so generated clients can inject it once.",
    "title": "Cloud Manager API",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/dashboard/summary": {
      "get": {
        "operationId": "dashboard_api_v1_dashboard_summary_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dashboard",
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/api/v1/instances": {
      "get": {
        "operationId": "list_instances_api_v1_instances_get",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "vm",
                    "container"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "power_state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "running",
                    "stopped",
                    "paused",
                    "unknown"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Power State"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_InstanceOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Instances",
        "tags": [
          "instances"
        ]
      },
      "post": {
        "operationId": "create_instance_api_v1_instances_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstanceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Instance",
        "tags": [
          "instances"
        ]
      }
    },
    "/api/v1/instances/{instance_id}": {
      "delete": {
        "operationId": "delete_instance_api_v1_instances__instance_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "instance_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Instance Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstanceDelete"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Instance",
        "tags": [
          "instances"
        ]
      },
      "get": {
        "operationId": "get_instance_api_v1_instances__instance_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "instance_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Instance Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Instance",
        "tags": [
          "instances"
        ]
      }
    },
...
    "/api/v1/instances/{instance_id}/{action}": {
      "post": {
        "operationId": "power_action_api_v1_instances__instance_id___action__post",
        "parameters": [
          {
            "in": "path",
            "name": "instance_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Instance Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "action",
            "required": true,
            "schema": {
              "enum": [
                "start",
                "stop",
                "shutdown",
                "restart"
              ],
              "title": "Action",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Power Action",
        "tags": [
          "instances"
        ]
      }
    },
    "/api/v1/jobs": {
      "get": {
        "operationId": "list_jobs_api_v1_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending",
                    "running",
                    "succeeded",
                    "failed",
                    "cancelled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "resource_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource Id"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "instance.power",
                    "instance.create",
                    "instance.delete"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_JobOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/v1/quotas": {
      "get": {
        "operationId": "tenant_quotas_api_v1_quotas_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QuotaLineOut"
                  },
                  "title": "Response Tenant Quotas Api V1 Quotas Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tenant Quotas",
        "tags": [
          "quotas"
        ]
      }
    },
...
  }
}